High resolution sounds were played with a significant delay due to the sound buffer size being too big.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@929 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/particles.c
@@ -858,7 +858,7 @@ if (gameOpts->render.smoke.bDisperse && !nType) {
 #if 0
 	decay = (float) sqrt (decay);
 #else
-	decay = pow (decay * decay * decay, 1.0 / 5.0);
+	decay = (float) pow (decay * decay * decay, 1.0 / 5.0);
 #endif
 	vOffset.p.x = f2fl (pParticle->nWidth) / decay;
 	vOffset.p.y = f2fl (pParticle->nHeight) / decay;
